fix bugs

+++ giz/data.py
@@ -42,7 +42,7 @@ class GIZDataset(Dataset):
 			y = self.map[y]
 
 			out.update(
-				{'target': torch.tensor(y, dtype=torch.float)}
+				{'target': torch.tensor(y, dtype=torch.long)}
 				)
 
 		return out
